requirements.txt
google-cloud-vision
google-cloud-firestore
typing
datetime
uuid